TNO-458: Combined view layout update
