+++ app/editor/src/features/content/combined-view/styled/CombinedView.tsx
@@ -5,12 +5,14 @@ export const CombinedView = styled(FormPage)`
   min-height: calc(100% - 500px);
   min-width: 95%;
   overflow: hidden;
+
   .reflex-element {
     padding-top: 0.5em;
-    height: 65%;
+    height: 100%;
   }
   .filter-area {
     background-color: #f5f5f5;
+    padding: 0.5em;
   }
   .content-filter {
     max-width: fit-content;
@@ -19,11 +21,9 @@ export const CombinedView = styled(FormPage)`
 
     .box {
       margin-top: 0.6em;
-      margin-left: 1em;
       border: solid 1px grey;
       border-radius: 0.25em;
-      padding: 1em;
-      max-width: fit-content;
+      padding: 0.5em;
     }
   }
 `;
