+++ app/editor/src/features/content/combined-view/styled/ListViewPanel.tsx
@@ -3,4 +3,7 @@ import { Col } from 'tno-core';
 
 export const ListViewPanel = styled(Col)`
   width: 150px;
+  .create-snippet {
+    width: fit-content;
+  }
 `;
